Move IntegerIndexType and TextIndexType into a common file
