--- lib/segment/src/data_types/integer_index.rs
@@ -1,20 +1,0 @@
-use schemars::JsonSchema;
-use serde::{Deserialize, Serialize};
-
-#[derive(Default, Debug, Deserialize, Serialize, JsonSchema, Clone, Copy, PartialEq, Hash, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum IntegerIndexType {
-    #[default]
-    Integer,
-}
-
-#[derive(Debug, Default, Deserialize, Serialize, JsonSchema, Clone, PartialEq, Hash, Eq)]
-#[serde(rename_all = "snake_case")]
-pub struct IntegerIndexParams {
-    // Required for OpenAPI schema without anonymous types, versus #[serde(tag = "type")]
-    pub r#type: IntegerIndexType,
-    /// If true - support direct lookups.
-    pub lookup: bool,
-    /// If true - support ranges filters.
-    pub range: bool,
-}
