+++ lib/segment/src/data_types/mod.rs
@@ -1,9 +1,8 @@
 pub mod groups;
-pub mod integer_index;
+pub mod index;
 pub mod named_vectors;
 pub mod order_by;
 pub mod primitive;
 pub mod query_context;
-pub mod text_index;
 pub mod tiny_map;
 pub mod vectors;
