+++ lib/segment/src/index/field_index/full_text_index/tests/mod.rs
@@ -3,7 +3,7 @@ use rstest::rstest;
 use tempfile::Builder;
 
 use crate::common::rocksdb_wrapper::open_db_with_existing_cf;
-use crate::data_types::text_index::{TextIndexParams, TextIndexType, TokenizerType};
+use crate::data_types::index::{TextIndexParams, TextIndexType, TokenizerType};
 use crate::index::field_index::full_text_index::text_index::FullTextIndex;
 use crate::index::field_index::{PayloadFieldIndex, ValueIndexer};
