+++ lib/segment/src/types.rs
@@ -25,9 +25,8 @@ use validator::{Validate, ValidationError, ValidationErrors};
 
 use crate::common::operation_error::{OperationError, OperationResult};
 use crate::common::utils::{self, MaybeOneOrMany, MultiValue};
-use crate::data_types::integer_index::IntegerIndexParams;
+use crate::data_types::index::{IntegerIndexParams, TextIndexParams};
 use crate::data_types::order_by::OrderValue;
-use crate::data_types::text_index::TextIndexParams;
 use crate::data_types::vectors::VectorStructInternal;
 use crate::index::field_index::CardinalityEstimation;
 use crate::index::sparse_index::sparse_index_config::SparseIndexConfig;
